+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * `SupervisorRestartError` explain message got enhanced with first detected error
   on a window of time (#65)
 
+* Introduce `NewDynSubtree` and `NewDynSubtreeWithNotifyStart` to create managed
+  subtrees with dynamic worker spawning capabilities. (#66)
+
 # v0.1.0 (breaking changes)
 
 * Deprecate `WithTolerance` worker option with `WithRestartTolerance` supervisor